+++ changelog.md
@@ -37,6 +37,7 @@ Codion Change Log
 - EntityTablePanel.ControlKeys.DECREMENT_SELECTION and INCREMENT_SELECTION keys changed from ALT-SHIFT to CTRL-SHIFT.
 - EntityApplicationPanel.Builder.LoginProvider removed, loginProvider() replaced with userSupplier().
 - EntityApplicationPanel.Builder.ConnectionProviderFactory removed, Builder.connectionProvider() added, related refactoring.
+- EntityApplicationPanel.Builder.defaultLoginUser(), automaticLoginUser() and userSupplier() renamed defaultUser(), user() and user() respectively.
 ### is.codion.tools.generator.model
 - DomainGeneratorModel now supports databases without schemas and users, as in, SQLite.
 ### is.codion.tools.generator.ui